Add TagFactsEditor to TagFacts plugin
This commit introduces the 'TagFactsEditor' which is added as a new module in the TagFacts plugin. The need for this is to expand the functionality of the TagFacts plugin and provide editing features. In detail, the new module includes its own build file 'TagFactsEditor.Build.cs' and the main file 'TagFactsEditor.cpp' that contains methods for startup and shutdown of the module. Furthermore, the addition extends dependencies in 'TagFacts.Build.cs' including 'GameplayTagsEditor' and reflects these changes in 'TagFacts.uplugin' with the new module and plugin details.