Remove GameplayTagsEditor dependency from TagFacts
The dependency on "GameplayTagsEditor" was removed from the plugin configurations in the TagFacts.uplugin and TagFacts.Build.cs, as well as from the module configurations in TagFactsEditor.Build.cs. This was done to decrease the complexity and potential error cases by reducing unnecessary dependencies. In addition, this change removes unnecessary code from FS_FactCustomization, simplifying the project and further reducing potential points of error.