Quest component is now a subsystem
This was done to achieve easier multi-world travel and easier to install and use.
The only reason to have it as a component is for multiplayer support, but I will not be doing any multiplayer and Flow's multiplayer documentation is too scarce at the moment.
This also means that there is no need for the FlowQuest interface, so it has been removed.